Rename checker sources to hpp
