+++ src/AbstractDoubleSumChecker.hpp
