+++ src/AndDoubleSumChecker.hpp
@@ -1,14 +1,14 @@
-#ifndef XORDOUBLESUMCHECKER_CPP
-#define XORDOUBLESUMCHECKER_CPP
+#ifndef ANDDOUBLESUMCHECKER_CPP
+#define ANDDOUBLESUMCHECKER_CPP
 
-#include "AbstractDoubleSumChecker.cpp"
+#include "AbstractDoubleSumChecker.hpp"
 #include <random>
 
-class XorDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
+class AndDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
     using T = long long;
 
     T func(T x, T y) override {
-        return x^y;
+        return x&y;
     }
 
     T SolveFasterAlgorithm(const std::vector<T>& A) override {
@@ -25,7 +25,7 @@ class XorDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
                     Y++;
                 }
             }
-            result+=(X*Y)<<k;
+            result+=(X*(X-1)/2)<<k;
         }
         return result;
     }
@@ -38,4 +38,4 @@ class XorDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
     }
 };
 
-#endif // XORDOUBLESUMCHECKER_CPP
+#endif // ANDDOUBLESUMCHECKER_CPP
