+++ src/MinDoubleSumChecker.hpp
@@ -1,15 +1,15 @@
-#ifndef DIFFABSDOUBLESUMCHECKER_CPP
-#define DIFFABSDOUBLESUMCHECKER_CPP
+#ifndef MINDOUBLESUMCHECKER_CPP
+#define MINDOUBLESUMCHECKER_CPP
 
-#include "AbstractDoubleSumChecker.cpp"
+#include "AbstractDoubleSumChecker.hpp"
 #include <random>
 #include <algorithm>
 
-class DiffAbsDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
+class MinDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
     using T = long long;
 
     T func(T x, T y) override {
-        return abs(x-y);
+        return std::min(x,y);
     }
 
     T SolveFasterAlgorithm(const std::vector<T>& A) override {
@@ -18,9 +18,10 @@ class DiffAbsDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
         std::vector<T> A_copy(N);
         std::copy(A.begin(),A.end(),A_copy.begin());
         std::sort(A_copy.begin(),A_copy.end());
-        for(int i=0;i<N;i++){
-            result+=(2*i+1-N)*A_copy[i];
+        for(int i=0;i<N;++i){
+            result+=(N-1-i)*A_copy[i];
         }
+
         return result;
     }
 
@@ -32,4 +33,4 @@ class DiffAbsDoubleSumChecker : public AbstractDoubleSumChecker<long long> {
     }
 };
 
-#endif // DIFFABSDOUBLESUMCHECKER_CPP
+#endif // MINDOUBLESUMCHECKER_CPP
