+++ tests/test_AddDoubleSumChecker.cpp
@@ -1,4 +1,4 @@
-#include "../src/AddDoubleSumChecker.cpp"
+#include "../src/AddDoubleSumChecker.hpp"
 #include <iostream>
 
 int main() {
